third_party/blink/web_tests/fast/forms/date/date-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type for "date" is correctly "date".
PASS input.type for "DATE" is correctly "date".
PASS input.type for " date " is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
